lesson_two/requirements.txt
num2words-0.5.10
transliterate-1.10.2